qwen3 cuda p2
